+++ changelog.md
@@ -11,3 +11,4 @@
 - quiver can now have arrows be inserted from off hand
 - blackboard undo button**
 - trapped presents will hide their names
+- you can now insert ropes into rope arrows from your inventory